Refactor DirectDruidClient, extract BufferInputStreamResponseHandler
